requirements.txt
flask==2.2.5
flask-cors==4.0.0
python-dotenv==0.21.0
groq>=0.9.0
flask-socketio==5.5.1
gunicorn==23.0.0
eventlet==0.38.2
google-generativeai==0.8.5
openai==0.27.0
requests==2.32.2
